% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/down.R
\name{downCorona}
\alias{downCorona}
\title{função que lê diretamente o gzip do servidor do brasil.io e retorna o respectivo csv}
\usage{
downCorona(file_url)
}
\description{
função que lê diretamente o gzip do servidor do brasil.io e retorna o respectivo csv
}
