% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/down.R
\name{pandate}
\alias{pandate}
\title{Função que recebe um dia a partir do dia zero ("20200225") a respectiva data no formato apropriado.}
\usage{
pandate(x)
}
\description{
Função que recebe um dia a partir do dia zero ("20200225") a respectiva data no formato apropriado.
}
