% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/down.R
\name{pegaCorona}
\alias{pegaCorona}
\title{pegaCorona(tipo = c("caso_full", "cart", "last_cases"), baixar = TRUE))
funcao que pega o arquivo de casos do brasil.io, faz um pequeno tratamento e retorna-o
retornos permitidos:
caso_full: retorna a contagem de casos de todos os municipios desde o dia 0
cart: obitos em cartório
last_cases: ultimo balanço divulgado, com latitude e longitude}
\usage{
pegaCorona(tipo = c("caso_full", "cart", "last_cases"), baixar = TRUE)
}
\description{
pegaCorona(tipo = c("caso_full", "cart", "last_cases"), baixar = TRUE))
funcao que pega o arquivo de casos do brasil.io, faz um pequeno tratamento e retorna-o
retornos permitidos:
caso_full: retorna a contagem de casos de todos os municipios desde o dia 0
cart: obitos em cartório
last_cases: ultimo balanço divulgado, com latitude e longitude
}
